- Replace arrays of int with arrays of short in the dynamics related class. (Some could be missings) - Small improvements in Interaction analysis (more coloration radioButton) - Draft DynamicalHierarchicalSimulation, synchronous updating should work at this point. Some problem with merging remains.

+++ src/fr/univmrs/tagc/GINsim/aRegGraph/GsRegulatoryAnimator.java
@@ -185,7 +185,7 @@ public class GsRegulatoryAnimator extends AbstractListModel implements GraphChan
      * 
      * @param state
      */
-    protected void colorizeGraph(int[] state) {
+    protected void colorizeGraph(short[] state) {
         if (state == null || state.length != nodeOrder.size()) {
             return;
         }
